[fx][inductor] Add statically_known_true utility for SymBool
This adds a function `statically_known_true` for `SymBool` that works
like inductor's `is_expr_static_and_true`. That is, it tries to simplify the
expression to a constant or returns `False` if it cannot be simplified.

This is useful in cases that can be optimized if the condition is met,
otherwise it doesn't effect correctness so we can avoid adding guards.

I also use this new function in inductor for `FakeTensorUpdater` and
`remove_noop_pass` which both generated unexpected guards previously.

ghstack-source-id: 1ace2ac843d3342f57b80361bb9dd6a551d29d50
Pull Request resolved: https://github.com/pytorch/pytorch/pull/117359